add weather to expanded.js
