+++ client/src/pages/Favorites.js
@@ -1,7 +1,7 @@
 import React, { useState, useCallback } from "react";
 import background from "../imgs/fav.jpeg";
 import PARKAPI from "../utils/auth/trailAPI/trailAPI";
-import { api_key } from "../api.json";
+
 import axios from "axios";
 
 const styles = {
@@ -25,42 +25,10 @@ const cardStyles = {
   textAlign: "center",
   backgroundColor: "#D3D3D3",
 };
-const weatherStyles = {
-  borderRadius: "100px",
-  color: "#1028F2",
-  fontWeight: "bold",
-  borderWidth: "50px",
-  display: "flex",
-  margin: "40px",
-  border: "5px black",
-};
 
 function Favorites() {
   const [park, setPark] = useState("");
-  const [weatherResponse, setWeatherResponse] = useState({});
   const [favoriteParks, setFavoriteParks] = useState([]);
-
-  const getParkWeather = () => {
-    const weatherURL = `http://api.openweathermap.org/data/2.5/forecast?q=${park}&units=imperial&APPID=${api_key}`;
-    fetch(weatherURL)
-      .then((res) => res.json())
-      .then((data) => {
-        console.log("weather:", data);
-        setWeatherResponse(data);
-      });
-  };
-  //   still need to figure out how the weatherurl will recieve it's data
-  const provideData = (event) => {
-    const val = event.target.value;
-    setPark(val);
-    console.log("park data:", park);
-  };
-
-  const onSubmit = useCallback((event) => {
-    event.preventDefault();
-    console.log("park:", park);
-    getParkWeather();
-  });
 
   React.useEffect(() => {
     PARKAPI.getParks().then(({ data }) => {
@@ -120,29 +88,6 @@ function Favorites() {
           </div>
         );
       })}
-
-      <div style={weatherStyles}>
-        {weatherResponse.list
-          ? weatherResponse.list.map((weatherItem, idx) => {
-              if (idx % 8 === 4) {
-                return (
-                  // forecast card
-                  <div key={idx} style={cardStyles}>
-                    <p>Date: {weatherResponse?.list[idx]?.dt_txt}</p>
-
-                    <p>
-                      Temperature: {weatherResponse?.list[idx]?.main?.temp} F
-                    </p>
-
-                    <p>
-                      Humidity: {weatherResponse?.list[idx]?.main?.humidity}%
-                    </p>
-                  </div>
-                );
-              }
-            })
-          : ""}
-      </div>
     </div>
   );
 }
